+++ packages/solid/src/components/button.tsx
@@ -71,6 +71,7 @@ export function UploadButton<TRouter extends FileRouter>(
     },
     onUploadError: $props.onUploadError,
     onUploadBegin: $props.onUploadBegin,
+    onBeforeUploadBegin: $props.onBeforeUploadBegin,
   });
 
   const fileInfo = () =>