+++ packages/solid/src/types.ts
@@ -19,6 +19,7 @@ export type UploadthingComponentProps<TRouter extends FileRouter> = {
      * @default (VERCEL_URL ?? window.location.origin) + "/api/uploadthing"
      */
     url?: string | URL;
+    onBeforeUploadBegin?: (files: File[]) => File[];
     multiple?: boolean;
   } & (undefined extends inferEndpointInput<TRouter[TEndpoint]>
     ? // eslint-disable-next-line @typescript-eslint/ban-types