+++ packages/uploadthing/src/internal/handler.ts
@@ -1,12 +1,15 @@
+import type { MimeType } from "@uploadthing/mime-types/db";
 import {
   generateUploadThingURL,
   getTypeFromFileName,
+  isObject,
   objectKeys,
   fillInputRouteConfig as parseAndExpandInputConfig,
   safeParseJSON,
   UploadThingError,
 } from "@uploadthing/shared";
 import type {
+  ContentDisposition,
   ExpandedRouteConfig,
   FileRouterInputKey,
   Json,
@@ -18,16 +21,35 @@ import { UPLOADTHING_VERSION } from "../constants";
 import { conditionalDevServer } from "./dev-hook";
 import { getParseFn } from "./parser";
 import { VALID_ACTION_TYPES } from "./types";
-import type { ActionType, FileRouter } from "./types";
+import type { ActionType, FileRouter, UTEvents } from "./types";
+
+/**
+ * Creates a wrapped fetch that will always forward a few headers to the server.
+ */
+const createUTFetch = (apiKey: string) => {
+  return async (endpoint: `/${string}`, payload: unknown) => {
+    const response = await fetch(generateUploadThingURL(endpoint), {
+      method: "POST",
+      body: JSON.stringify(payload),
+      headers: {
+        "Content-Type": "application/json",
+        "x-uploadthing-api-key": apiKey,
+        "x-uploadthing-version": UPLOADTHING_VERSION,
+      },
+    });
+
+    return response;
+  };
+};
 
 const fileCountLimitHit = (
-  files: string[],
+  files: { name: string }[],
   routeConfig: ExpandedRouteConfig,
 ) => {
   const counts: Record<string, number> = {};
 
   files.forEach((file) => {
-    const type = getTypeFromFileName(file, objectKeys(routeConfig));
+    const type = getTypeFromFileName(file.name, objectKeys(routeConfig));
 
     if (!counts[type]) {
       counts[type] = 1;
@@ -77,10 +99,16 @@ const getHeader = (req: RequestLike, key: string) => {
   return req.headers[key];
 };
 
-type UploadThingResponse = {
-  presignedUrl: { url: string; fields: Record<string, string> }; // ripped type from S3 package
-  name: string;
+export type UploadThingResponse = {
+  presignedUrls: string[];
   key: string;
+  pollingUrl: string;
+  uploadId: string;
+  fileName: string;
+  fileType: MimeType;
+  contentDisposition: ContentDisposition;
+  chunkCount: number;
+  chunkSize: number;
 }[];
 
 export const buildRequestHandler = <TRouter extends FileRouter>(
@@ -203,12 +231,11 @@ export const buildRequestHandler = <TRouter extends FileRouter>(
       });
     }
 
+    const utFetch = createUTFetch(preferredOrEnvSecret);
+
     switch (actionType) {
       case "upload": {
-        const maybeInput = await safeParseJSON<{
-          files: string[];
-          input: Json;
-        }>(req);
+        const maybeInput = await safeParseJSON<UTEvents["upload"]>(req);
 
         if (maybeInput instanceof Error) {
           return new UploadThingError({
@@ -253,11 +280,19 @@ export const buildRequestHandler = <TRouter extends FileRouter>(
         }
 
         // Validate without Zod (for now)
-        if (!Array.isArray(files) || !files.every((f) => typeof f === "string"))
+        if (
+          !Array.isArray(files) ||
+          !files.every(
+            (f) =>
+              isObject(f) &&
+              typeof f.name === "string" &&
+              typeof f.size === "number",
+          )
+        )
           return new UploadThingError({
             code: "BAD_REQUEST",
-            message: "Files must be a string array",
-            cause: `Expected files to be of type 'string[]', got '${JSON.stringify(
+            message: "Files must be an array of objects with name and size",
+            cause: `Expected files to be of type '{name:string, size:number}[]', got '${JSON.stringify(
               files,
             )}'`,
           });
@@ -300,26 +335,15 @@ export const buildRequestHandler = <TRouter extends FileRouter>(
 
         const callbackUrl = url.origin + url.pathname;
 
-        const uploadthingApiResponse = await fetch(
-          generateUploadThingURL("/api/prepareUpload"),
-          {
-            method: "POST",
-            body: JSON.stringify({
-              files: files,
+        const uploadthingApiResponse = await utFetch("/api/prepareUpload", {
+          files: files,
 
-              routeConfig: parsedConfig,
+          routeConfig: parsedConfig,
 
-              metadata,
-              callbackUrl: config?.callbackUrl ?? callbackUrl,
-              callbackSlug: slug,
-            }),
-            headers: {
-              "Content-Type": "application/json",
-              "x-uploadthing-api-key": preferredOrEnvSecret,
-              "x-uploadthing-version": UPLOADTHING_VERSION,
-            },
-          },
-        );
+          metadata,
+          callbackUrl: config?.callbackUrl ?? callbackUrl,
+          callbackSlug: slug,
+        });
 
         // This is when we send the response back to the user's form so they can submit the files
         const parsedResponse = await safeParseJSON<UploadThingResponse>(
@@ -343,12 +367,18 @@ export const buildRequestHandler = <TRouter extends FileRouter>(
           }
         }
 
-        return { body: parsedResponse, status: 200 };
+        return {
+          body: parsedResponse.map((x) => ({
+            ...x,
+            pollingUrl: generateUploadThingURL(`/api/pollUpload/${x.key}`),
+          })),
+          status: 200,
+        };
       }
-      case "failure": {
-        const maybeReqBody = await safeParseJSON<{
-          fileKey: string;
-        }>(req);
+      case "multipart-complete": {
+        const maybeReqBody = await safeParseJSON<
+          UTEvents["multipart-complete"]
+        >(req);
         if (maybeReqBody instanceof Error) {
           return new UploadThingError({
             code: "BAD_REQUEST",
@@ -356,23 +386,37 @@ export const buildRequestHandler = <TRouter extends FileRouter>(
             cause: maybeReqBody,
           });
         }
-        const { fileKey } = maybeReqBody;
+
+        const completeRes = await utFetch("/api/completeMultipart", {
+          fileKey: maybeReqBody.fileKey,
+          uploadId: maybeReqBody.uploadId,
+          etags: maybeReqBody.etags,
+        });
+        if (!completeRes.ok) {
+          return new UploadThingError({
+            code: "UPLOAD_FAILED",
+            message: "Failed to complete multipart upload",
+          });
+        }
+
+        return { status: 200 };
+      }
+      case "failure": {
+        const maybeReqBody = await safeParseJSON<UTEvents["failure"]>(req);
+        if (maybeReqBody instanceof Error) {
+          return new UploadThingError({
+            code: "BAD_REQUEST",
+            message: "Invalid request body",
+            cause: maybeReqBody,
+          });
+        }
+        const { fileKey, uploadId } = maybeReqBody;
 
         // Tell uploadthing to mark the upload as failed
-        const uploadthingApiResponse = await fetch(
-          generateUploadThingURL("/api/failureCallback"),
-          {
-            method: "POST",
-            body: JSON.stringify({
-              fileKey,
-            }),
-            headers: {
-              "Content-Type": "application/json",
-              "x-uploadthing-api-key": preferredOrEnvSecret,
-              "x-uploadthing-version": UPLOADTHING_VERSION,
-            },
-          },
-        );
+        const uploadthingApiResponse = await utFetch("/api/failureCallback", {
+          fileKey,
+          uploadId,
+        });
 
         if (!uploadthingApiResponse.ok) {
           console.error("[UT] failed to mark upload as failed");