+++ packages/uploadthing/src/sdk/utils.ts
@@ -7,7 +7,8 @@ import {
   UploadThingError,
 } from "@uploadthing/shared";
 
-import { maybeParseResponseXML } from "../internal/s3-error-parser";
+import { uploadPart } from "../internal/multi-part";
+import type { UTEvents } from "../server";
 
 export function guardServerOnly() {
   if (typeof window !== "undefined") {
@@ -76,7 +77,7 @@ export const uploadFilesInternal = async (
   });
 
   if (!res.ok) {
-    const error = await UploadThingError.fromResponse(res as Response);
+    const error = await UploadThingError.fromResponse(res);
     throw error;
   }
 
@@ -84,26 +85,29 @@ export const uploadFilesInternal = async (
   const json = await res.json<
     | {
         data: {
-          presignedUrl: string; // url to post to
-          fields: Record<string, string>;
+          presignedUrls: string[];
           key: string;
-          fileUrl: string; // the final url of the file after upload
+          fileUrl: string;
+          fileType: string;
+          uploadId: string;
+          chunkSize: number;
+          chunkCount: number;
         }[];
       }
     | { error: string }
   >();
 
   if ("error" in json) {
-    const error = await UploadThingError.fromResponse(clonedRes as Response);
+    const error = await UploadThingError.fromResponse(clonedRes);
     throw error;
   }
 
-  // Upload each file to S3
+  // Upload each file to S3 in chunks using multi-part uploads
   const uploads = await Promise.allSettled(
     data.files.map(async (file, i) => {
-      const { presignedUrl, fields, key, fileUrl } = json.data[i];
+      const { presignedUrls, key, fileUrl, uploadId, chunkSize } = json.data[i];
 
-      if (!presignedUrl || !fields) {
+      if (!presignedUrls || !Array.isArray(presignedUrls)) {
         throw new UploadThingError({
           code: "URL_GENERATION_FAILED",
           message: "Failed to generate presigned URL",
@@ -111,56 +115,41 @@ export const uploadFilesInternal = async (
         });
       }
 
-      const formData = new FormData();
-      formData.append("Content-Type", file.type);
-      Object.entries(fields).forEach(([key, value]) => {
-        formData.append(key, value);
-      });
+      const etags = await Promise.all(
+        presignedUrls.map(async (url, index) => {
+          const offset = chunkSize * index;
+          const end = Math.min(offset + chunkSize, file.size);
+          const chunk = file.slice(offset, end);
 
-      formData.append(
-        "file",
-        // Handles case when there is no file name
-        file.name
-          ? (file as File)
-          : Object.assign(file as File, { name: "unnamed-blob" }),
+          const etag = await uploadPart({
+            fetch: opts.fetch,
+            url,
+            chunk: chunk as Blob,
+            contentDisposition: data.contentDisposition,
+            contentType: file.type,
+            fileName: file.name,
+            maxRetries: 10,
+            key,
+            utRequestHeaders: opts.utRequestHeaders,
+          });
+
+          return { tag: etag, partNumber: index + 1 };
+        }),
       );
 
-      // Do S3 upload
-      const s3res = await opts.fetch(presignedUrl, {
+      // Complete multipart upload
+      await opts.fetch(generateUploadThingURL("/api/completeMultipart"), {
         method: "POST",
-        body: formData,
-        headers: new Headers({
-          Accept: "application/xml",
-        }),
+        body: JSON.stringify({
+          fileKey: key,
+          uploadId,
+          etags,
+        } satisfies UTEvents["multipart-complete"]),
+        headers: opts.utRequestHeaders,
       });
 
-      if (!s3res.ok) {
-        // tell uploadthing infra server that upload failed
-        await opts.fetch(generateUploadThingURL("/api/failureCallback"), {
-          method: "POST",
-          body: JSON.stringify({
-            fileKey: fields.key,
-          }),
-          headers: opts.utRequestHeaders,
-        });
-
-        const text = await s3res.text();
-        const parsed = maybeParseResponseXML(text);
-        if (parsed?.message) {
-          throw new UploadThingError({
-            code: "UPLOAD_FAILED",
-            message: parsed.message,
-          });
-        }
-        throw new UploadThingError({
-          code: "UPLOAD_FAILED",
-          message: "Failed to upload file to storage provider",
-          cause: s3res,
-        });
-      }
-
       // Poll for file to be available
-      await pollForFileData(key);
+      await pollForFileData(generateUploadThingURL(`/api/pollUpload/${key}`));
 
       return {
         key,