+++ packages/uploadthing/CHANGELOG.md
@@ -1,5 +1,18 @@
 # uploadthing
 
+## 5.7.4
+
+- Updated dependencies
+  [[`eb5f96d`](https://github.com/pingdotgg/uploadthing/commit/eb5f96dc06a81ecb4b1f7ee3d0ba259ebdfee7d1)]:
+  - @uploadthing/mime-types@0.2.2
+  - @uploadthing/shared@5.2.7
+
+## 5.7.3
+
+- Updated dependencies
+  [[`da43651`](https://github.com/pingdotgg/uploadthing/commit/da436516a9c1e30268878016a9c1dc930bc384f6)]:
+  - @uploadthing/shared@5.2.6
+
 ## 5.7.2
 
 ### Patch Changes